Moved app from containers into components.
